Danmaku part now indenpendent, named wpfDanmakulib
