+++ src/OhMyDanmaku/OhMyDanmaku.csproj
@@ -61,6 +61,7 @@
     <Compile Include="Settings.xaml.cs">
       <DependentUpon>Settings.xaml</DependentUpon>
     </Compile>
+    <Compile Include="wpfDanmakulib.cs" />
     <Page Include="Audit.xaml">
       <SubType>Designer</SubType>
       <Generator>MSBuild:Compile</Generator>
